requirements.txt
dataclasses
protobuf
grpcio
grpcio-tools